+++ bindings/python/Cargo.toml
@@ -9,4 +9,4 @@ crate-type = ["cdylib"]
 
 [dependencies]
 pyo3 = { version = "0.20.0", features = ["extension-module"] }
-codelist-rs = { path = "../rust/codelist-rs" }
+codelist-rs = { path = "../../rust/codelist-rs" }
