--- bindings/python/src/codelists/bindings.py
@@ -1,1 +1,0 @@
-#
